feat: business layer <> domain layer 로직 위치 변경

+++ src/business/user/user.business.ts
@@ -1,13 +1,13 @@
-import { BadRequestException, Injectable, Logger, NotFoundException } from '@nestjs/common';
+import { BadRequestException, ConflictException, Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { GetUser } from '@src/common/decorators/get-user.decorator';
 import { PaginationQueryDto } from '@src/common/dtos/pagination/pagination-query.dto';
 import { PaginatedResponseDto } from '@src/common/dtos/pagination/pagination-response.dto';
-import { UserEntity } from '@src/domain/user/entities/user.entity';
-import { UserDomainService } from '@src/domain/user/user.service';
 import { plainToInstance } from 'class-transformer';
 import { UpdateUserDto } from './dto/requests/update-lams-user.dto';
 import { UserResponseDto } from './dto/responses/user-response.dto';
 import { UserRole } from '@src/domain/user/enum/user.enum';
+import { UserDomainService } from '@src/domain/user/services/user-domain.service';
+import { LamsUserEntity } from '@src/domain/user/entities/lams-user.entity';
 
 /**
  * 사용자 비즈니스 서비스
@@ -33,7 +33,7 @@ export class UserBusinessService {
     /**
      * 사용자 프로필 조회 - 단순 조회: 로그 불필요
      */
-    async getProfile(@GetUser() user: UserEntity): Promise<UserResponseDto> {
+    async getProfile(@GetUser() user: LamsUserEntity): Promise<UserResponseDto> {
         const response = plainToInstance(UserResponseDto, user);
         return response;
     }
@@ -54,6 +54,12 @@ export class UserBusinessService {
      * 사용자 생성 - 비즈니스 크리티컬: 로그 필요
      */
     async createUser(username: string, email: string, password: string): Promise<UserResponseDto> {
+        // 이메일 중복 확인
+        const existingUser = await this.userDomainService.findUserByEmail(email);
+        if (existingUser) {
+            throw new ConflictException('이미 존재하는 이메일입니다.');
+        }
+
         const user = await this.userDomainService.createUser(username, email, password);
         const response = {
             userId: user.userId,
@@ -72,8 +78,23 @@ export class UserBusinessService {
     /**
      * 사용자 정보 업데이트 - 비즈니스 크리티컬: 로그 필요
      */
-    async updateUser(userId: string, updateUserDto: UpdateUserDto): Promise<UserResponseDto> {
-        const result = await this.userDomainService.updateUser(userId, updateUserDto);
+    async updateUser(userId: string, dto: UpdateUserDto): Promise<UserResponseDto> {
+        const user = await this.userDomainService.findUserById(userId);
+        if (!user) {
+            throw new NotFoundException('사용자를 찾을 수 없습니다.');
+        }
+
+        const updateData: Partial<LamsUserEntity> = {};
+        if (dto.username) updateData.username = dto.username;
+        if (dto.password) {
+            const isValidate = user.validatePassword(dto.password);
+
+            if (!isValidate) {
+                updateData.password = user.updateHashedPassword(dto.password);
+            }
+        }
+
+        const result = await this.userDomainService.updateUser(userId, updateData);
         const response = plainToInstance(UserResponseDto, result);
         this.logger.log(`사용자 정보 업데이트 완료: ${result.email}`);
         return response;
@@ -97,7 +118,9 @@ export class UserBusinessService {
             throw new BadRequestException('현재 비밀번호가 올바르지 않습니다.');
         }
 
-        const result = await this.userDomainService.changePassword(userId, newPassword);
+        const hashedPassword = user.updateHashedPassword(newPassword);
+
+        const result = await this.userDomainService.changePassword(userId, hashedPassword);
         const response = plainToInstance(UserResponseDto, result);
         this.logger.log(`비밀번호 변경 완료: ${result.email}`);
         return response;
@@ -107,6 +130,12 @@ export class UserBusinessService {
      * 사용자 삭제 - 비즈니스 크리티컬: 로그 필요
      */
     async deleteUser(userId: string): Promise<boolean> {
+        const user = await this.userDomainService.findUserById(userId);
+
+        if (!user) {
+            throw new NotFoundException('사용자를 찾을 수 없습니다.');
+        }
+
         const result = await this.userDomainService.deleteUser(userId);
         this.logger.log(`사용자 삭제 완료: ${userId}`);
         return result;
@@ -116,11 +145,14 @@ export class UserBusinessService {
      * 사용자 역할 업데이트 - 비즈니스 크리티컬: 로그 필요
      */
     async updateUserRole(userId: string, roles: UserRole[]): Promise<UserResponseDto> {
-        if (!userId || !roles || roles.length === 0) {
-            throw new BadRequestException('유효하지 않은 사용자 ID 또는 역할 정보입니다.');
+        const user = await this.userDomainService.findUserById(userId);
+        if (!user) {
+            throw new NotFoundException('사용자를 찾을 수 없습니다.');
         }
 
-        const result = await this.userDomainService.updateUserRole(userId, roles);
+        const roleStrings = roles.map((role) => role.toString());
+
+        const result = await this.userDomainService.updateUserRole(userId, roleStrings);
         const response = plainToInstance(UserResponseDto, result);
         this.logger.log(`사용자 역할 업데이트 완료: ${result.email}, 새 역할: ${roles.join(', ')}`);
         return response;
